requirements.txt
flask
opencv-python
mediapipe
numpy
scikit-learn
pillow
soundfile
librosa
matplotlib
pandas